Start using a new request reading scheme.
This is intended to be an implementation closely tied with the spec and
using strong typing rigorously, instead of the typical string typing.

The code for this is fairly verbose, but is distinctly faster than the
old approach.

The old scheme (for reading the request line) is still there as
``request.old.rs``, but will be removed very soon. One thing it has
which the new scheme doesn't (and can't conveniently, at present—TODO)
is tests.