Fix everything up. It almost all works again now.
All known headers still get dropped. This causes problems, naturally.
But fear not: this too shall pass.

Actually, the server will only successfully respond to HTTP/1.0 requests
at present, because it requires ``request.host`` to not be ``None``,
which this is making it be. HTTP/1.1 requests will promptly get a terse
``500 Bad Request``.