Use a more robust "header value finished" check.
This won't affect the particular case it's used in now, but will affect
other header values that I'm working on.

This *is* now a destructive function (as noted in the method docs), but
I believe that will normally be what I require. For other cases, there
is still ``self.state == Finished``; it may be safer not to have this in
a method, hence its present removal.