Add configure to the README
This pull request just adds `.configure` to the Getting started section; I know it's basic, but for someone just moving into the language, it is a barrier to entry.

Instructions are also given for rust-openssl to "run its `configure` and `make`", with a reference to rust-http's `configure` - this just localizes that instruction to where it is needed.