Simplify the build system.
Remove the `configure` script and the Makefile. This disables optional SSL,
but allows us to switch to Cargo completely. Code generation is now in its own package,
and is called from Cargo through `prebuild.sh`.

The `doc` folder was renamed to `documentation` to avoid deletion when
running `cargo clean`. The examples were moved to the root so they get built when
running `cargo test`.